Fix freeze on some workout steps
Some notes in workout files are not handled correctly by the FIT SDK and
cause the program to go into an endless loop when saving the merged fit
file.

This commit fixes the issue by removing the notes from the workout
steps. While this is a loss of information, it is not that crucial for
analysis as the workout steps with their duration and targets are still
present in the merged file.